% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net2blendfuncs.R
\name{add_missing_vertex}
\alias{add_missing_vertex}
\title{Add missing vertices to a network.}
\usage{
add_missing_vertex(net, allvertex, attrlist = list())
}
\arguments{
\item{net}{igraph network object}

\item{allvertex}{Vector of edges as generated by \code{find_all_vertex}}

\item{attrlist}{list of attributes to be given to the new vertices, where the name
of each list item is the name of an vertex attribute.}
}
\value{
Returns a network with the vertices of \code{allvertex} that are not
    present in \code{net} added. The new vertices will have the attributes
    provided by \code{attrlist}.
}
\description{
\code{add_missing_vertices} Add edges to a network to ensure consistency.
}
\examples{
data(examplenet1)
data(examplenet2)
netlist=list(examplenet1,examplnet2)#list of nets
alledges=find_all_edges(netlist)#get all edges
allvertices=find_all_vertex(netlist)#get all vertices
#add missing vertices to examplenet1
examplenet1=add_missing_vertex(examplenet1,allvertices)
}
