% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net2blendfuncs.R
\name{find_all_edges}
\alias{find_all_edges}
\title{Find all possible edges in a list of networks}
\usage{
find_all_edges(allnets)
}
\arguments{
\item{allnets}{list of igraph network objects}
}
\value{
Returns a dataframe of all edges that exist within \code{allnets}.
    This can be used as an argument in \code{add_missing_edges} to ensure
    that all edges are correctly generated for animation in Blender.
}
\description{
\code{find_all_edges} Find all edges that exist in a list of networks.
}
\examples{
data(examplenet1)
data(examplenet2)
netlist=list(examplenet1,examplnet2)
find_all_edges(netlist)
}
