% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net2blendfuncs.R
\name{findmaxlength}
\alias{findmaxlength}
\title{Find longest edge in a layout}
\usage{
findmaxlength(net, layout)
}
\arguments{
\item{net}{igraph network object}

\item{layout}{Matrix or dataframe of the same number of rows as vertices
in the networks, where first column is vertex x coordinates
and second column is vertex y coordinates. Z coordinates can also be
provided.}
}
\value{
Returns the length of the longest edge in a given layout. This can
    then be used as the \code{maxlength} argument in \code{net2blend} to keep
    curved edges consistent over multiple networks.
}
\description{
\code{findmaxlength} Calculate longest edge in a given layout
}
