% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net2blendfuncs.R
\name{get_edge_names}
\alias{get_edge_names}
\title{Get edge names}
\usage{
get_edge_names(net)
}
\arguments{
\item{net}{igraph network object}
}
\value{
Returns a vector of edge names in the format "vertex1_vertex2"
}
\description{
\code{get_edge_names} Utility function to assign each edge a name.
}
