% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net2blendfuncs.R
\name{add_missing_edges}
\alias{add_missing_edges}
\title{Add missing edges to a network.}
\usage{
add_missing_edges(net, alledges = NULL, attrlist = list())
}
\arguments{
\item{net}{igraph network object}

\item{alledges}{Dataframe of edges as generated by \code{find_all_edges}}

\item{attrlist}{list of attributes to be given to the new edges, where the name
of each list item is the name of an edge attribute.}
}
\value{
Returns a network with the edges of \code{alledges} that are not
    present in \code{net} added. The new edges will have the attributes
    provided by \code{attrlist}.
}
\description{
\code{add_missing_edges} Add edges to a network to ensure consistency.
}
