% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net2blendfuncs.R
\name{find_all_vertex}
\alias{find_all_vertex}
\title{Find all possible edges in a list of networks}
\usage{
find_all_vertex(allnets)
}
\arguments{
\item{allnets}{list of igraph network objects}
}
\value{
Returns a vector of all vertices that exist within \code{allnets}.
    This can be used as an argument in \code{add_missing_vertex} to ensure
    that all edges are correctly generated for animation in Blender.
}
\description{
\code{find_all_vertex} Find all edges that exist in a list of networks.
}
